brieffy-backend/requirements.txt
fastapi
uvicorn
google-auth
python-dotenv
pyjwt
pydantic
pydantic[email]
email-validator
pymongo
selenium
beautifulsoup4
lxml 
html5lib
openai
requests
retry